Разбиение кода на модули, подключение модулей через AMD
