+++ src/main.js
@@ -1,7 +1,3 @@
 'use strict';
 
-require('./resizer');
-require('./upload');
-require('./pictures');
-
-
+define(['./upload', './resizer', './pictures'], function() {});
