+++ src/resizer.js
@@ -1,6 +1,6 @@
 'use strict';
 
-(function() {
+define(function() {
   /**
    * @constructor
    * @param {string} image
@@ -355,4 +355,4 @@
   };
 
   window.Resizer = Resizer;
-})();
+});
